refactor: move set-piece definitions to data/biomes/corporate-pieces.js
Set-piece definitions are biome content, not infrastructure. Each biome
will have its own pieces file.

- js/core/network/set-pieces.js: now just instantiate() + typedefs (233 lines)
- data/biomes/corporate-pieces.js: all 20 corporate set-pieces + 5 atomics
- js/core/network/atomics.js: removed (merged into corporate-pieces)
- All imports updated across networks, biome, mini-network, and tests

523 tests pass, both generated and hand-crafted networks work.
